requirements.txt
# Core dependencies
aiohttp==3.8.5
asyncio==3.4.3
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1

# Security and networking
requests==2.31.0
urllib3==2.0.4
certifi==2023.7.22
cryptography==41.0.4
pyopenssl==23.2.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Database
sqlite3

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6
tabulate==0.9.0
tldextract==5.1.1

# DNS and networking
dnspython==2.4.2

# Report generation
weasyprint==60.1
jinja2==3.1.2

# Development and testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Additional utilities
click==8.1.7
rich==13.5.2
typer==0.9.0
